Add files via upload
This is a presentation made by me where I analyzed the data provided by a America based startup Dognition which helps people to know better about their dogs by giving them some tests. The objective of this analysis is to determine business strategy that would increase the numbers of tests users complete on Dognition website by 10 % on a month to month basis compared to the same month of last year.